moved Sandwiches example into Factory Method solution
